pyproject.toml
[tool.poetry]
name = "esmjax"
version = "0.1.0"
description = "Inference of the ESM2 Protein Language Models on JAX/Flax"
authors = ["Irhum Shafkat <13271155+irhum@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
flax = "^0.6.0"
einops = "^0.4.1"
jaxtyping = "^0.2.7"
torch = "^1.12.1"
tokenizers = "^0.13.0"


[tool.poetry.group.dev.dependencies]
jax = {extras = ["cpu"], version = "^0.3.19"}
fair-esm = "^1.0.2"
pytest = "^7.1.3"
jupyter = "^1.0.0"
ufmt = "^2.0.1"
seaborn = "^0.12.0"
black = {extras = ["jupyter"], version = "^22.8.0"}
biopython = "^1.79"


[[tool.poetry.source]]
name = "jax[tpu]"
url = "https://storage.googleapis.com/jax-releases/libtpu_releases.html"
default = false
secondary = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
